pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "anthropic>=0.36.0",
    "black>=24.0.0",
    "chromadb>=0.5.0",
    "cohere>=5.0.0",
    "faiss-cpu>=1.8.0",
    "fastapi>=0.110.0",
    "groq>=0.11.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-huggingface>=0.1.0",
    "langchain-ollama>=0.2.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    "mypy>=1.10.0",
    "numpy>=1.24.0",
    "openai>=1.50.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "python-dotenv>=1.0.0",
    "python-json-logger>=2.0.0",
    "pyyaml>=6.0.0",
    "rank-bm25>=0.2.2",
    "requests>=2.31.0",
    "ruff>=0.5.0",
    "sentence-transformers>=3.0.0",
    "sse-starlette>=2.0.0",
    "structlog>=24.0.0",
    "tenacity>=8.2.0",
    "tiktoken>=0.7.0",
    "unstructured>=0.14.0",
    "uvicorn>=0.30.0",
]
